+++ src/discounts/components/VoucherList/VoucherList.tsx
@@ -9,6 +9,7 @@ import TableCellHeader from "@saleor/components/TableCellHeader";
 import TableHead from "@saleor/components/TableHead";
 import TablePagination from "@saleor/components/TablePagination";
 import { VoucherListUrlSortField } from "@saleor/discounts/urls";
+import { canBeSorted } from "@saleor/discounts/views/VoucherList/sort";
 import { maybe, renderCollection } from "@saleor/misc";
 import { makeStyles } from "@saleor/theme";
 import { ChannelProps, ListActions, ListProps, SortPage } from "@saleor/types";
@@ -130,6 +131,9 @@ const VoucherList: React.FC<VoucherListProps> = props => {
           }
           textAlign="right"
           onClick={() => onSort(VoucherListUrlSortField.minSpent)}
+          disabled={
+            !canBeSorted(VoucherListUrlSortField.minSpent, !!selectedChannelId)
+          }
           className={classes.colMinSpent}
         >
           <FormattedMessage
@@ -175,6 +179,9 @@ const VoucherList: React.FC<VoucherListProps> = props => {
           }
           textAlign="right"
           onClick={() => onSort(VoucherListUrlSortField.value)}
+          disabled={
+            !canBeSorted(VoucherListUrlSortField.minSpent, !!selectedChannelId)
+          }
           className={classes.colValue}
         >
           <FormattedMessage